Moved gl_windows.hpp into cross_platform/windows_wrapper.hpp.
