+++ oss_src/cross_platform/windows_wrapper.hpp
