+++ oss_src/fileio/libhdfs_shim.cpp
@@ -18,7 +18,7 @@
 #ifndef _WIN32
 #include <dlfcn.h>
 #else
-#include <process/gl_windows.hpp>
+#include <cross_platform/windows_wrapper.hpp>
 #include <util/syserr_reporting.hpp>
 #endif
 extern  "C" {
