+++ oss_src/fileio/oss_webstor/sysutils.cpp
@@ -25,7 +25,7 @@
 
 #ifdef _WIN32
 #define WIN32_LEAN_AND_MEAN
-#include <process/gl_windows.hpp>
+#include <cross_platform/windows_wrapper.hpp>
 #include <ws2tcpip.h>
 #include <mstcpip.h>
 #else  // !_WIN32
