+++ oss_src/minipsutil/minipsutil.c
@@ -35,7 +35,7 @@ int32_t kill_process(int32_t pid);
 /*                                Windows                                 */
 /*                                                                        */
 /**************************************************************************/
-#include <process/gl_windows.hpp>
+#include <cross_platform/windows_wrapper.hpp>
 #include <Psapi.h>
 
 typedef BOOL (WINAPI *LPFN_GLPI)(
