+++ oss_src/util/syserr_reporting.cpp
@@ -8,7 +8,7 @@
 #include <util/syserr_reporting.hpp>
 #include <iostream>
 #ifdef _WIN32
-#include <process/gl_windows.hpp>
+#include <cross_platform/windows_wrapper.hpp>
 #else
 #include <errno.h>
 #include <string.h>
